+++ src/main/java/org/cojen/maker/ConstantsRegistry.java
@@ -126,6 +126,7 @@ public abstract class ConstantsRegistry {
      *
      * @param name unused
      * @param type unused
+     * @param slot if bit 31 is set, then also remove the constant
      */
     public static Object find(MethodHandles.Lookup lookup, String name, Class<?> type, int slot) {
         if ((lookup.lookupModes() & ACCESS_MODE) == 0) {
@@ -157,12 +158,45 @@ public abstract class ConstantsRegistry {
             }
         }
 
-        if (value instanceof Entries) {
-            value = ((Entries) value).mValues[slot];
+        if (value == null) {
+            return null;
         }
 
-        if (value == null) {
-            throw new IllegalStateException();
+        if (value instanceof Entries) {
+            var entries = (Entries) value;
+            synchronized (entries) {
+                value = entries.mValues[slot & Integer.MAX_VALUE];
+                if (value == null) {
+                    return null;
+                }
+                int size = entries.mSize;
+                if (size > 1) {
+                    if (slot < 0) {
+                        entries.mValues[slot & Integer.MAX_VALUE] = null;
+                        entries.mSize = size - 1;
+                    }
+                    return value;
+                }
+            }
+        }
+
+        if (slot < 0) {
+            if (loader instanceof ClassInjector.Group) {
+                var group = (ClassInjector.Group) loader;
+                synchronized (group) {
+                    group.mConstants.remove(clazz);
+                    if (group.mConstants.isEmpty()) {
+                        group.mConstants = null;
+                    }
+                }
+            } else {
+                synchronized (registry) {
+                    registry.mConstants.remove(clazz);
+                    if (registry.mConstants.isEmpty()) {
+                        registry.mConstants = null;
+                    }
+                }
+            }
         }
 
         return value;