+++ src/test/java/org/cojen/maker/CondyTest.java
@@ -333,6 +333,56 @@ public class CondyTest {
     }
 
     @Test
+    public void sneaky() throws Exception {
+        // Try to steal a complex constant, to verify some of the security features. When given
+        // enough access, the constant can be stolen. This isn't a feature, but just an
+        // artifact of the current implementation.
+
+        ClassMaker cm = ClassMaker.begin(null, MethodHandles.lookup());
+        cm.addField(byte[].class, "test").public_().static_().final_();
+
+        MethodMaker mm = cm.addClinit();
+        var const0 = new byte[] {1,2,3};
+        mm.field("test").setExact(const0);
+
+        var lookup = cm.finishHidden();
+        var clazz = lookup.lookupClass();
+
+        // Class hasn't been initialized yet. Try to steal via the backdoor.
+
+        int slot = 1 << 31;
+
+        assertNull(ConstantsRegistry.find(MethodHandles.lookup(), "_", null, slot));
+
+        try {
+            ConstantsRegistry.find(MethodHandles.lookup().in(clazz), "_", null, slot);
+            fail();
+        } catch (IllegalStateException e) {
+            // Doesn't have private access.
+        }
+
+        try {
+            ConstantsRegistry.find
+                (lookup.dropLookupMode(MethodHandles.Lookup.PRIVATE), "_", null, slot);
+            fail();
+        } catch (IllegalStateException e) {
+            // Doesn't have private access.
+        }
+
+        if (Runtime.version().feature() < 15) {
+            // Hidden classes aren't really supported, and so private lookup isn't available.
+            return;
+        }
+
+        // Works when given full permission.
+        Object const1 = ConstantsRegistry.find(lookup, "_", null, slot);
+        assertEquals(const0, const1);
+
+        // Stolen!
+        assertNull(clazz.getField("test").get(null));
+    }
+
+    @Test
     public void classDesc() throws Exception {
         ClassDesc cd0 = ClassDesc.of("java.lang.String").arrayType();
         ClassDesc cd1 = ClassDesc.ofDescriptor("I");